Centralize quantum seed handling
