+++ vybn/__init__.py
@@ -1,0 +1,5 @@
+"""Vybn package."""
+
+from .quantum_seed import seed_rng
+
+__all__ = ["seed_rng"]
